Move index case to new OapIndexQuerySuite for better organization.
