+++ src/test/scala/org/apache/spark/sql/execution/datasources/oap/OapDDLSuite.scala
@@ -123,20 +123,5 @@ class OapDDLSuite extends QueryTest with SharedSQLContext with BeforeAndAfterEac
       new Configuration()).globStatus(new Path(path,
       "oap_partition_table/b=2/c=c2/*.index")).length != 0)
   }
-
-  test("index integrity") {
-      val data: Seq[(Int, String)] = scala.util.Random.shuffle(1 to 300).map{
-                                      i => (i, s"this is test $i") }.toSeq
-      data.toDF("key", "value").createOrReplaceTempView("t")
-      sql("insert overwrite table oap_test_1 select * from t")
-      sql("create sindex index1 on oap_test_1 (a) using bitmap")
-
-      val dfwithIdx = sql("SELECT * FROM oap_test_1 WHERE a > 8 and a <= 200")
-      sql("drop sindex index1 on oap_test_1")
-      val dfWithoutIdx = sql("SELECT * FROM oap_test_1 WHERE a > 8 and a <= 200")
-      val dfOriginal = sql("SELECT * FROM t WHERE key > 8 and key <= 200")
-      assert(dfWithoutIdx.count == dfwithIdx.count)
-      assert(dfWithoutIdx.count == dfOriginal.count)
-  }
 }
